sec5-b16
The purpose of our project is to develop a system that can automatically detect cancer
from the blood cell images. This system uses a convolution network that inputs a blood
cell images and outputs whether the cell is infected with cancer or not. The appearance
of cancer in blood cell images is often vague, can overlap with other diagnoses, and can
mimic many other benign abnormalities. These discrepancies cause considerable
variability among medical personnel in the diagnosis of cancer. Automated detection
of cancer from blood cell images at the level of expert medical personnel would not
only have tremendous benefit in clinical settings, it would also be invaluable in delivery
of health care to populations with inadequate access to diagnostic imaging specialists.