Heart disease prediction using bio inspired algorithms
The heart disease Prediction algorithm is a user support and on-line consultation project. The health care industry produces a huge amount of data. This data is not always made use to the full extent and is often underutilized. Using this huge amount of data, a disease can be detected, predicted or even cured. A huge threat to humankind is caused by diseases like heart disease, cancer, tumor and Alzheimer’s disease.This model can then be used to predict future medical data using some machine learning algorithms. The accuracy of the model using each of the algorithms is calculated. Then the one with good accuracy is taken as the model for predicting the heart disease. According to result, the patient will contact doctor consequently for any treatment.